refactor setup import responsibilities
